aws-lambda/requirements.txt
selenium
requests
gspread
PyDrive
google-api-python-client
oauth2client
python-dateutil
pykakasi